requirements..txt
streamlit
pandas
numpy
joblib
google-genai
opencv-python-headless
easyocr